requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=py36_0
av=7.0.1=py36h66a6ed3_1
blas=1.0=mkl
blinker=1.4=py_1
brotlipy=0.7.0=py36h8c4c3a4_1000
bzip2=1.0.8=h7b6447c_0
c-ares=1.16.1=h516909a_0
ca-certificates=2020.6.20=hecda079_0
cachetools=4.1.1=py_0
cairo=1.16.0=hcf35c78_1003
certifi=2020.6.20=py36h9f0ad1d_0
cffi=1.14.1=py36h0ff685e_0
chardet=3.0.4=py36h9f0ad1d_1006
click=7.1.2=pyh9f0ad1d_0
cryptography=3.0=py36h45558ae_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py_2
dbus=1.13.6=he372182_0
expat=2.2.9=he1b5a44_2
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=h86ecdb6_1001
freeglut=3.0.0=hf484d3e_5
freetype=2.10.2=h5ab3b9f_0
future=0.18.2=py36h9f0ad1d_1
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.2.1=h516909a_2
gitdb=4.0.5=py_0
gitpython=3.1.7=py_0
glib=2.65.0=h6f030ca_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.13=h79a8f9a_0
google-auth=1.20.1=py_0
google-auth-oauthlib=0.4.1=py_2
graphite2=1.3.14=h23475e2_0
grpcio=1.31.0=py36h769ab6c_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.5=nompi_h3c11f04_1104
icu=64.2=he1b5a44_1
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=py_0
importlib-metadata=1.7.0=py36h9f0ad1d_0
intel-openmp=2020.1=217
jasper=1.900.1=h07fcdf6_1006
joblib=0.16.0=py_0
jpeg=9d=h516909a_0
kiwisolver=1.2.0=py36hdb11119_0
lame=3.100=h14c3975_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
libclang=9.0.1=default_hde54327_0
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libiconv=1.16=h516909a_0
liblapack=3.8.0=16_mkl
liblapacke=3.8.0=16_mkl
libllvm9=9.0.1=he513fc3_1
libopencv=4.2.0=py36_3
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0=h8b12597_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libvpx=1.7.0=h439df22_0
libwebp=1.0.2=h56121f0_5
libxcb=1.14=h7b6447c_0
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
lz4-c=1.9.2=he6710b0_1
markdown=3.2.2=py_0
matplotlib=3.3.0=1
matplotlib-base=3.3.0=py36h5ffbc53_1
mkl=2020.1=217
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.1.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=h1bed415_1002
ninja=1.10.0=py36hfd86e86_0
nspr=4.27=he1b5a44_1
nss=3.55=he751ad9_0
numpy=1.19.1=py36hbc911f0_0
numpy-base=1.19.1=py36hfa32c7d_0
nvidia-dali-cuda100=0.24.0=pypi_0
oauthlib=3.0.1=py_0
olefile=0.46=py36_0
opencv=4.2.0=py36_3
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1g=h516909a_1
pandas=1.0.5=py36h0573a6f_0
pcre=8.44=he6710b0_0
pillow=7.2.0=py36hb39fc2d_0
pip=20.2.1=py36_0
pixman=0.38.0=h516909a_1003
protobuf=3.13.0=py36h831f99a_0
py-opencv=4.2.0=py36h95af2a2_3
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pyjwt=1.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pysocks=1.7.1=py36h9f0ad1d_1
python=3.6.11=h425cb1d_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.6.0=py3.6_cuda10.2.89_cudnn7.6.5_0
pytorch-lightning=0.8.5=py_0
pytz=2020.1=py_0
pyyaml=5.3.1=py36h8c4c3a4_0
qt=5.12.5=hd8c4c69_1
readline=8.0=h7b6447c_0
requests=2.24.0=pyh9f0ad1d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.6=pyh9f0ad1d_0
scikit-learn=0.23.1=py36h423224d_0
scipy=1.5.0=py36h0b6359f_0
seaborn=0.10.1=py_0
setuptools=49.2.1=py36_0
six=1.15.0=py_0
smmap=3.0.4=pyh9f0ad1d_0
sqlite=3.32.3=h62c20be_0
tensorboard=2.3.0=py_0
tensorboard-plugin-wit=1.6.0=pyh9f0ad1d_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
torchvision=0.7.0=py36_cu102
tornado=6.0.4=py36h8c4c3a4_1
tqdm=4.48.2=pyh9f0ad1d_0
urllib3=1.25.10=py_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py36_0
x264=1!152.20180806=h14c3975_0
xlrd=1.2.0=py36_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.11=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
